Python_advanced/import_module_package/requirements.txt
python-dotenv~=1.0.0
telebot==0.0.5
